==> ms_occ/ORTWS_OccModel/For Participants/MultispeciesOccupancy/Full Example/00_Rsetup.R <==
## Run this script first to install the needed packages before running any of the example scripts

install.packages("reshape",dependencies=TRUE)
install.packages("R2jags",dependencies=TRUE)
install.packages("ggplot2",dependencies=TRUE)

## load the packages to make sure they loaded properly
library(reshape)
library(R2jags)
library(ggplot2)
